requirements.txt
Django==3.1.3
djangorestframework==3.12.2
Telethon==1.17.5
pyyaml==5.3.1
django-admin-rangefilter==0.6.3
